task: update arrow dependency to 16.1.0
the current version of grpc in homebrew, 1.62.2, seems to be
incompatible with arrow 15.0.1 due to a dependency on protobuf::upb.
This issue goes away (as far as I can tell) for arrow 16.1.0. I have not
triaged the root cause, but updating to arrow 16.1.0 seems to be the
easy thing for now.